Add server option to hide intel on other nations if you have contact
When server option 'contact_intel' is disabled, having contact (but no
embassy) will not reveal the other nation's government, gold and diplomatic
relations. The client may show Anarchy as the government and all relations
as War, since it's not aware of the server option. Having an embassy still
reveals all intel as before. The default is that the option is enabled,
giving the traditional behaviour.