Add /showtimeout and /fixtimeout, fix /syncturn to work right
Server command /showtimeout shows the time left of current turn (phase)
and other such times. /fixtimeout changes the time remaining on the current
turn only.

/syncturn changes the time remaining on the current turn so as to synchronise
the game to a global schedule based on the timeout setting and the game
port number. This patch fixes it to use game.tinfo.seconds_to_phasedone
correctly and adds the 'show' argument.

'/syncturn N' sets the turn to end on the next sync point, so that there's
at least N x timeout time left of current turn. So with '/syncturn .5', it's
the one at least 11.5 h away for 23 h timeouts. '/syncturn show' shows the
time of the next sync point. '/syncturn' without an argument is the same
as '/syncturn 0.8'.

The schedule of sync points comes from the unix time (seconds since start of
1970) modulo game timeout length, modified by the game number.  E.g.  for a
game with port number ending with 9, and a timeout of 23 h, one sync point
is on 2020-10-23 8:00:00 UTC, the next is 23 h from that (2020-10-24 7:00:00
UTC) and so on.